refactor(flui-layer): Mythos Step 6 -- extract LayerBounds trait to bounds.rs
Extract the `LayerBounds` trait + 11 production impls from
`crates/flui-layer/src/layer/mod.rs` to a new
`crates/flui-layer/src/layer/bounds.rs` file. This is the last of the
mechanical extractions called out by the verdict's Section 6 module layout
for `layer/mod.rs` (the others -- generated dispatch in dispatch.rs --
landed in Mythos Step 4).

LOC delta:

- `layer/mod.rs`: 656 -> 584 LOC (-72 LOC).
- `layer/bounds.rs`: new, 103 LOC (including module doc + trait doc +
  inherent re-export pattern for each impl).
- Net: +31 LOC due to the module docstring + per-impl spacing; the
  pre-extraction trait + impls were ~78 LOC of tightly-packed code with
  no module-level documentation.

The 11 layer types that implement `LayerBounds` (Canvas, ClipRect,
ClipRRect, ClipPath, ClipSuperellipse, ShaderMask, BackdropFilter, Texture,
PlatformView, Leader, AnnotatedRegion) are unchanged -- the trait still
forwards `LayerBounds::bounds(&self)` to each type's inherent `bounds()`
method. The 7 layer types that intentionally do NOT implement the trait
(Offset, Transform, Opacity, ColorFilter, ImageFilter, Follower,
PerformanceOverlay) are documented in the new module's docstring.

mod.rs structure after this step:

- Module docstring (~40 LOC)
- mod declarations + pub use re-exports (~50 LOC)
- `Layer` enum definition (~60 LOC)
- impl Layer { bounds, needs_compositing, is_clip, is_linking, is_opaque }
  (~110 LOC; only the multi-arm-match methods stay -- the 57 single-variant
  is_*/as_*/as_*_mut methods went to dispatch.rs in Step 4)
- 19 `impl From<XxxLayer> for Layer` (~95 LOC; candidates for a future macro
  pass, deferred -- not blocking the chain)
- `dispatch::gen_layer_accessors!` macro invocation (~25 LOC)
- inline tests (~120 LOC)

The verdict targeted ~300 LOC of production code in mod.rs; at 584 LOC
total (~344 LOC excluding tests + module docs) the file is still above
target. The remaining surplus is the 19 `impl From<XxxLayer>`
implementations -- a candidate for a `gen_layer_from_impls!` macro in a
future pass, recorded as Outstanding Refactor in U12's
`crates/flui-layer/ARCHITECTURE.md`.

Files touched:

- Create: `crates/flui-layer/src/layer/bounds.rs`.
- Modify: `crates/flui-layer/src/layer/mod.rs`:
  * Add `mod bounds;` + `pub use bounds::LayerBounds;`.
  * Delete the `// ========== LAYER BOUNDS TRAIT ==========` section
    (lines 492-566 in the pre-extract file).
  * The `pub use` ordering puts bounds before dispatch so the trait surface
    is reachable from `layer/mod.rs` re-exports as
    `pub use layer::LayerBounds`.

Verification:

- `cargo check --workspace`: clean.
- `cargo test -p flui-layer --lib`: 263 passed, 0 failed.
- `wc -l crates/flui-layer/src/layer/{mod,bounds}.rs`: 584 + 103 = 687.

Plan: docs/plans/2026-05-20-002-feat-flui-layer-mythos-redesign-plan.md U6.
